Struct ethers_signers::Wallet

pub struct Wallet<D: DigestSigner<ProxyDigest<Sha256>, RecoverableSignature>> { /* fields omitted */ }

An Ethereum private-public key pair which can be used for signing messages.

Examples

Signing and Verifying a message

The wallet can be used to produce ECDSA Signature objects, which can be then verified. Note that this uses hash_message under the hood which will prefix the message being hashed with the Ethereum Signed Message domain separator.

use ethers_core::rand::thread_rng;
use ethers_signers::{LocalWallet, Signer};

let wallet = LocalWallet::new(&mut thread_rng());

// Optionally, the wallet's chain id can be set, in order to use EIP-155
// replay protection with different chains
let wallet = wallet.with_chain_id(1337u64);

// The wallet can be used to sign messages
let message = b"hello";
let signature = wallet.sign_message(message).await?;
assert_eq!(signature.recover(&message[..]).unwrap(), wallet.address());

Implementations

impl Wallet<SigningKey>

pub fn new_keystore<P, R, S>(
dir: P,
rng: &mut R,
password: S
) -> Result<(Self, String), WalletError> where
P: AsRef<Path>,
R: Rng + CryptoRng,
S: AsRef<[u8]>,

Creates a new random encrypted JSON with the provided password and stores it in the provided directory. Returns a tuple (Wallet, String) of the wallet instance for the keystore with its random UUID.

pub fn decrypt_keystore<P, S>(
keypath: P,
password: S
) -> Result<Self, WalletError> where
P: AsRef<Path>,
S: AsRef<[u8]>,

Decrypts an encrypted JSON from the provided path to construct a Wallet instance

pub fn new<R: Rng + CryptoRng>(rng: &mut R) -> Self

Creates a new random keypair seeded with the provided RNG

impl<D: DigestSigner<ProxyDigest<Sha256>, RecoverableSignature>> Wallet<D>

pub fn sign_transaction_sync(&self, tx: &TypedTransaction) -> Signature

Synchronously signs the provided transaction.

pub fn sign_hash(&self, hash: H256, eip155: bool) -> Signature

Signs the provided hash and proceeds to normalize the v value of the signature with EIP-155 if the flag is set to true.

pub fn signer(&self) -> &D

Gets the wallet’s signer