Crate ethers_signers

Provides a unified interface for locally signing transactions.

Re-exports

pub use coins_bip39;

Structs

Ledger

A Ledger Ethereum App.

MnemonicBuilder

Represents a structure that can resolve into a Wallet<SigningKey>.

Trezor

A Trezor Ethereum App.

Wallet

An Ethereum private-public key pair which can be used for signing messages.

Enums

HDPath

Ledger wallet type

LedgerError

Error when using the Ledger transport

TrezorError

Error when using the Trezor transport

TrezorHDPath

Trezor wallet type

WalletError

Error thrown by the Wallet module

Traits

Signer

Trait for signing transactions and messages

Functions

to_eip155_v

Applies EIP155

Type Definitions

LocalWallet

A wallet instantiated with a locally stored private key